update: drawer navigation is implemented
changes:
1. initial stack navigation is changed with drawer navigation
2. seperate `/home` route is created to contain landing page and article
   page
3. stack navigation is used in `/home` route
4. packages updated
