+++ app/_layout.tsx
@@ -1,32 +1,15 @@
 import { GestureHandlerRootView } from 'react-native-gesture-handler';
 import { Drawer } from 'expo-router/drawer';
 
-export default function Layout() {
 
-  // const navigation = useNavigation()
-  // return (
-  //     <Stack 
-  //     screenOptions={{
-  //       title: 'Dev Community',
-  //       headerRight: ()=> <TouchableOpacity onPress={()=>alert('not implemented')} hitSlop={46}><Icon source="tag" size={36} color={'black'}/></TouchableOpacity>
-  //     }}>
-  //       <Stack.Screen 
-  //       name="[article]" 
-  //       options={{
-  //         title: 'Dev Community',
-  //         headerLeft: () => (
-  //           <TouchableOpacity hitSlop={25} onPress={()=> navigation.goBack()}><Icon source={'arrow-left'} size={25}/></TouchableOpacity>
-  //         ),
-  //         }}
-  //        />
-  //     </Stack>
-  // );
+export default function Layout() {
   return (
     <GestureHandlerRootView style={{ flex: 1 }}>
     <Drawer>
-      <Drawer.Screen name='index'
+      <Drawer.Screen name='home'
+      
         options={{
-          title: 'Home'
+          title: 'Dev Community'
         }}
       />
       <Drawer.Screen name='about'
@@ -34,6 +17,20 @@ export default function Layout() {
           title: 'About'
         }}
       />
+      <Drawer.Screen
+      name='index'
+      options={{
+        drawerItemStyle: {display: 'none'},
+        title: 'Dev Community'
+      }}
+      />
+      <Drawer.Screen
+      name='+not-found'
+      options={{
+        drawerItemStyle: {display: 'none'}
+      }}
+      />
+
     </Drawer>
     </GestureHandlerRootView>
   );
