+++ app/home/_layout.tsx
@@ -1,0 +1,8 @@
+import { Stack } from 'expo-router'
+
+export default function Layout() {
+    return (<Stack>
+        <Stack.Screen name='index' options={{headerShown: false}}/>
+        <Stack.Screen name='[article]' options={{headerShown: false}}/>
+    </Stack>);
+}
